backend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blis==0.7.11
bs4==0.0.2
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cssselect==1.2.0
cymem==2.0.8
Django==5.0.6
django-cors-headers==4.3.1
django-filter==24.3
django-redis==5.4.0
django-stubs==5.0.4
django-stubs-ext==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.16.1
frozenlist==1.4.1
googlesearch-python==1.2.4
idna==3.7
inflection==0.5.1
jieba3k==0.35.1
Jinja2==3.1.4
joblib==1.4.2
langcodes==3.4.0
langdetect==1.0.9
language_data==1.2.0
lxml==5.2.2
lxml_html_clean==0.2.2
marisa-trie==1.1.1
MarkupSafe==2.1.5
multidict==6.0.5
murmurhash==1.0.10
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.4
packaging==24.0
pillow==10.4.0
preshed==3.0.9
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
redis==5.0.4
regex==2024.5.15
requests==2.31.0
requests-file==2.1.0
scikit-learn==1.5.2
scipy==1.14.1
sgmllib3k==1.0.0
six==1.16.0
smart-open==6.4.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.0
srsly==2.4.8
thinc==8.2.3
threadpoolctl==3.5.0
tinysegmenter==0.3
tldextract==5.1.2
tomli==2.0.1
tqdm==4.66.4
typer==0.9.4
types-PyYAML==6.0.12.20240808
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
wasabi==1.1.2
weasel==0.3.4
yarl==1.9.4
